Improve performance when skipping ESO and GPL tests
A lot of the time spent while skipping shader object and graphics
pipeline library test is spent obtaining the list of extension
properties for the physical device. This happens, for example, every
time the main utility functions are used:

  * checkShaderObjectRequirements
  * checkPipelineConstructionRequirements

These functions are written to be as generic as possible, so they
receive an instance interface and a physical device, and are used
sometimes with the context default device and some other times with
custom devices.

Internally, they use enumerateDeviceExtensionProperties, which is a
function that returns a vector of VkExtensionProperties. Building this
vector is time consuming, as is requesting the extension list to the
implementation over and over again.

The solution implemented here is providing a new variant of this
function that caches results by instance interface and physical device,
and selectively replacing some calls to the old function with calls to
the result-caching functions. In hot code paths, like the utility
functions mentioned above, this can save minutes of runtime.

When skipping ESO and GPL tests, we've also detected a lot of time is
spent writing to the test log file, specially in the writeEscaped
function. Replacing one printf call inside it with fputs appears to make
it slightly faster, saving some seconds overall.

With the changes proposed in this commit on my development laptop,
skipping the pipeline ESO tests goes from taking 6m30s to 1m30s.

No test results should be affected by this change, and no tests that
were previously supported should turn into unsupported and vice-versa.

Affects:
dEQP-VK.*

Components: Framework
VK-GL-CTS issue: 4704

Change-Id: Ie8846fbae17132c9bf460cf6e12253b07a5cffe0