Destroy all device helpers in dynamic state tests
The device helpers used in dEQP-VK.dynamic_state.* are not destroyed for
ESO test subgroups. As these are global variables using unique_ptr, the
objects are destroyed in any case at the end of the program, but that
may happen at a bad unpredictable moment, after everything else is
already gone, including the Vulkan library or the instance that was used
to create the devices.

Affects:
dEQP-VK.dynamic_state.*

VK-GL-CTS issue: 4714
Components: Vulkan

Change-Id: I1e107f3d1e20e29a214e4c24d63524f1266166ae