Add option to make test log more compact
This will remove redundant lines from the log output and all additional information that is not used to verify the test results, like <Text>, <Number>, <ImageSet>, etc.
The compact log can be obtained by passing --deqp-log-compact=enable

This is a non-breaking change and the full log with all the information can still be obtained with that option disabled.

The compact log is less than 20% of the size of the full log, and it is faster to parse.

Default output will be the same as usual, while the --deqp-log-compact=enable option will cause the following log to be emitted.
Example (for all tests):

<TestCaseResult CasePath="dEQP-VK.pipeline.monolithic.sampler.view_type.1d.format.r16g16_sint.min_reduce.comp_identity_zero_zero_zero.max">
 <Result StatusCode="NotSupported">Unsupported format for linear filtering: VK_FORMAT_R16G16_SINT</Result>
</TestCaseResult>
<TestCaseResult CasePath="dEQP-VK.pipeline.monolithic.sampler.view_type.1d_array.format.r4g4_unorm_pack8.mag_reduce.comp_r_g_b_a.average">
 <Result StatusCode="Pass">Result image matches reference</Result>
</TestCaseResult>

Additionally, these tests:

dEQP-VK.info.*
dEQP-VK.api.info.*
dEQP-VK.api.version_check.*

will emit full log output, including <text>, <number>, <imageset>, etc. for inspection purpose.

VK-GL-CTS issue: 4717

This does not affect any test, but the log output for all of them.

Change-Id: I3f6ea82694ae47b5c3c2a96a15ee16a5458f41c6