Take into account subpass dependencies in render pass wrapper
Shader objects pipeline tests use a render pass wrapper that allows the
code to work with both classic pipelines and shader objects. This
wrapper internally handles layout transitions from the initial layout to
the layouts used in the different subpasses and transitions from the
last used layout to the final layout specified for attachments.

Passes and subpasses are converted into dynamic rendering commands, and
layout transitions are inserted as barriers between these dynamic render
passes.

Translating classic render passes into dynamic rendering is a nontrivial
problem. For example, Mesa has 2000+ lines of code to deal with this and
the solution is not perfect nor complete (the code has several TODO
comments), but for CTS we do not need to be that thorough and adding
something that works for the cases that use the wrapper is enough.

One notable step that was missing was handling subpass dependencies and
inserting them as memory barriers between subpasses. This commit adds
some bits of code to do that.

In addition, the access and stage flags for external subpass
dependencies (either as the source or destination subpasses) are now
ORed with other access and stage flags in the initial layout
transitions and the final layout transitions.

Affects (for the reported issue):
dEQP-VK.pipeline.shader_object_*.*.sample_locations_ext.*

Affects (in general):
dEQP-VK.pipeline.shader_object_*

VK-GL-CTS issue: 4720
Components: Vulkan

Change-Id: Ife219e0c1708a188cf59b7210933dcf231f2a2bd