Make sure the stencil test passes if load op is "none"
In the load_store_op_none tests, some tests use a stencil attachment
with a load op of "none". In those cases, the contents of the stencil
buffer are undefined and, if the stencil test is enabled, it may not
pass unless the compare op is "always".

Affects:
dEQP-VK.*.load_store_op_none.*

Components: Vulkan
VK-GL-CTS issue: 4790

Change-Id: I43a1593ffb4bcaed9292e8dfe11f9135e6f4c7ef