Add a VUID skip list
To help catch regressions while documenting the list of known validation
failures, this change adds a simple skip list. In the process, it
switches CTS's use of the deprecated VK_EXT_debug_report extension to
the VK_EXT_debug_utils extension.

The actual list of known failures is incomplete and should be filled in
as existing failures are discovered.

Components: Vulkan
VK-GL-CTS issue: 4766

Change-Id: Ia751418de05875b7ead42c37bfea44be814d3cc7