Fix EGL render tests for rgb10a2 and rgb10a0
This commit makes EGL render tests work for rgb10a2 formats, by forcing
alpha values to be either 0 or 1, which is needed so that the fuzzy
comparison algorithm used is able to properly work without having to
increase the single floating point error threshold so much that it
doesn't make sense for components other than alpha. See the comments in
the code for more details.

In addition, this commits adds a new format to getColorFormat to avoid
asserts on rgb10a0.

Affects:
dEQP-EGL.functional.render.*

Components: AOSP
VK-GL-CTS issue: 4823

Change-Id: I95631e7b96aae6a9459159510604b4732c60b346