Fix nearest_edge cases for RGB10 surfaces
These tests are asserting when the surface is RGB10 but lacks an alpha
component.

Note the format translates to RGBA+UNORM_INT_1010102_REV as that seems
to be the format used by implementations for these surfaces. Alpha is
not important in this case, and it's set to 1.0 when preparing the
textures, which matches the returned surfaces.

Affects:
KHR-*.nearest_edge.*

Components: OpenGL
VK-GL-CTS issue: 4822

Change-Id: I23443638a0c1659d1c7a9654457c088da02ac581