Add sparse memory rebinding tests
There was no coverage for binding, unbinding and rebinding memory back
to the resource. These new tests excercise this functionality by
binding memory objects fully and then back - partially to the resources.

Image tests are not covering every possible format as that is not
imporant for  the testing of the rebinding functionality and it is
not to unnecessary extend the run-time of the CTS as the other formats
are already covered by other sparse binding tests.

New tests:
dEQP-VK.sparse_resources.buffer.transfer.rebind.*
dEQP-VK.sparse_resources.image_rebind.*

Components: Vulkan

VK-GL-CTS issue: 4853

Change-Id: I29ddac92b4bd825412394563004459acdff3c525