Fix the precision loss issue in native d16
The reference pixel is generated by comparing a value that is the same
as the depth read by the gpu with the value of depth of type float
generated in test. However, the accuracy of UNORM values in memory does
not match that of float values, especially d16, which varies greatly. In
the case of some edge values, the values of the reference pixels may be
different because the comparison results may be different due to the
processing of precision.
This commit introduces a temporary 1x1 dimensional texture2D image to
store the same reference depth as the depth buffer format used by the
GPU. In this way, we can make the depth comparison in test consistent
with GPU hardware in both precision and format, thus avoiding the
previous accuracy loss problem.

Affected tests:
dEQP-EGL.functional.image.create*
dEQP-EGL.functional.image.render_multiple_contexts*
dEQP-EGL.functional.image.modify*

VK-GL-CTS issue: 4800

Components: EGL
Change-Id: Ie6d51923a021bd21e0b15672b64f41f246c177b6