Fix incorrect barrier for unused attachment
In the current tests the unused attachment is cleared in a first
command buffer :

  vkCmdPipelineBarrier(unusedImage, UNDEFINED, TRANSFER_DST_OPTIMAL);
  vkCmdClearColor(unusedImage);
  vkCmdPipelineBarrier(unusedImage, TRANSFER_DST_OPTIMAL, COLOR_ATTACHMENT_OPTIMAL);

Then in a second buffer the image is again transfered from the
UNDEFINED layout :

  vkCmdPipelineBarrier(unusedImage, UNDEFINED, COLOR_ATTACHMENT_OPTIMAL);
  vkCmdBeginRendering();
  ...
  vkCmdEndRendering();

This second transfer from the UNDEFINED layout is loosing the clear
writes on Intel/Mesa. We can just skip that transition since the
previous command buffer already put the image in the right layout.

Components: Vulkan

VK-GL-CTS issue: 4519

Affects:
dEQP-VK.dynamic_rendering.*.unused_attachment.*

Change-Id: Iad634858a09007d93c0d1a25b96b702a537330a9