Always generate 4 component border colours
The sampler border colour swizzle tests were only generating as many
border colour components as the format being tested, which leads to
problems for (at least) the A8_UNORM tests. Switch to generating all 4
components for every test. This works around the A8 problem and (very
slightly) simplifies the code and increases coverage.

The A8 tests were generating only a single component, which meant that
the border colour was always (C, 0, 0, 0). Since the border colour is
taken from the A component the correct result was always 0 regardless of
swizzle. With this fix the results will be correctly randomised.

Components: Vulkan
Affects: dEQP-VK.pipeline.*.sampler.border_swizzle.*
Change-Id: I5f7f779d11ece892bc7362a7a279adbebb134e9a