Improve buffer view testing for A8 formats
These tests generated a reference colour or (r,g,b,1) with r, g and b
varying between pixels. However, because this was being written to a
pixel buffer access with format of A8_UNORM all the random data was
being discarded and the buffer was filled with nothing but 1s. Add some
random data to the alpha channel as well.

Components: Vulkan
Affects: dEQP-VK.api.buffer_view.access.*

Change-Id: Id6d07f967eadf5a40bda0c05030e0289f968b8fa