Decrease max.no.concurrent devices in DeviceGroup
MAX_CONCURRENT_DEVICES is set to 32. It's also used for DeviceGroup, but
it's not taken into account that multiple devices are allocated in a
DeviceGroup.
This fix divides MAX_CONCURRENT_DEVICES by the number of devices in the
device group.

Affects:
dEQP-VK.api.object_management.max_concurrent.device_group

Components: Vulkan

VK-GL-CTS issue: 4978

Change-Id: Ibe17092670b1f5127b0041720d65a358c5ea1ffb