Add fallback verification method for AV1 filmgrain
Implementations are allowed to diverge from the process specified in
the AV1 standard for film grain synthesis, and still be considered a
conformant implementation. The CTS currently performs a bit-exact
check for the process given by the standard, failing implementations
that technically are compliant.

The fallback's method is to create two decoding sessions, one with
film grain applied as usual, and a second with film grain forcibly
disabled. PSNR is computed on the outputs of both sessions, and
checked to be within a reasonable range. This is a non-principled
comparison.

Components: Vulkan

VK-GL-CTS issue: 5016

Affects:

dEQP-VK.video.*

Change-Id: Ib998008002736836612ccde5eeaf15edf3c6560e