Consider layer index when retrieving decoded frame
In certain cases, a layered image may be in use for the DPB and
retrieving the data for the decoded frame requires using the layer
index. This is a no-op for the case when the DPB is represented by
individual images.

Components: Vulkan

VK-GL-CTS issue: 5008

Affected tests:
dEQP-VK.video.encode.*

Change-Id: I4733f52ab703423d56a414c46f3c506dc2536dbf