Fix wsi maintenance1 release image tests
This patch aims to fix the wsi maintenance1 release image tests that
were breaking a few Vulkan specifications.

1) The test breaks the vkAcquireNextImage spec by not supplying a
valid synchronisation primitive when presentIndex != 0.

This is fixed by supplying and waiting on a VkFence in the image
acquires, even if the image will not be presented.

2) The test breaks the vkReleaseSwapchainImages spec by potentially
asking for 0 images to be released, which can happen if we only
acquired one image and that image will be presented.

This is fixed by skipping calling vkReleaseSwapchainImages if the
number of images to release are 0.

3) The tests break the VkSurfacePresentModeEXT spec by querying for
the capabilities of a presentation mode that was not reported as
supported by vkGetPhysicalDeviceSurfacePresentModesKHR.

This is fixed by moving the call to getPerPresentSurfaceCapabilities()
after getPhysicalDeviceSurfacePresentModes().

Affects:
dEQP-VK.wsi.*.maintenance1.release_images.*

Components: Vulkan

VK-GL-CTS issue: 5180

Change-Id: Id9d3d118c938e3258e6bbbdf1e84fdd92af3d993