Remove 'decor' specialisation in SPIR-V tests
The parameterised switch in the 8bit_storage tests was switching between
a decoration and a storage class, depending on whether the tests were
for uniform or storage buffers. This meant that it couldn't ever have
worked to use the template to specialise the tests this way and it
recently led to a validation error breaking the tests. This fixes the
underlying confusion so that hopefully it's easier to avoid this
happening again.

Component: Vulkan
Affects: dEQP-VK.spirv_assembly.instruction.*.8bit_storage.*

Change-Id: I6de79a19fda6b22ee751912a92eee3a476d5218d