Fix pointer error from real bool type
Having replaced deBool with real bool it is no longer guaranteed to be
the size of an int. This means that we need to be careful when copying
from a shader buffer where bools are 4 bytes.

In this example, from findLSBZero, 'bool comparison' was a single byte,
placed before 'int32_t res', meaning that the first 3 bytes of 'res'
were overwritten by 'comparison'.

Components: OpenGLES
Affects: dEQP-GLES31.functional.shaders.builtin_functions.uniform.*

Change-Id: I6c9c0512f97598ac4a3bca5164e1e6400ec7c321