Add missing dependencies on extensions for video encode tests
All video encode tests require the presence of VK_KHR_video_encode_queue
but this extension is not specified in the list of required device
extensions, so this change adds it to the list.

Additionally, the tests added for VK_KHR_video_maintenance1 have missed
requiring the codec-specific extensions, and this change addresses this
bug as well.

Components: Vulkan

VK-GL-CTS issue: 5143

Affected tests:
dEQP-VK.video.encode.*

Change-Id: Iaff9c6b27e14f96578f5bbd0767e5e5d399a2970