Run clang-format + remove de* types
This runs the reformat script format_all.sh
that syoussefi@ developed.

The same reformat script is ran in the remote tracking branch
and the reformat changes have landed in the remote tracking branch:
https://googleplex-android.git.corp.google.com/platform/external/deqp/+/b8dfc435c451c61e3564f51bbe5ccdfe9c8ce226

This CL completes the step 2) as part of the task to
get android external/deqp repo in sync with
remote tracking branch goog/upstream-vulkan-cts-1.3.8 after
the reformat change:

1) merge remote tracking branch goog/upstream-vulkan-cts-1.3.8
up to the commit 7d1b3a:
https://googleplex-android.git.corp.google.com/platform/external/deqp/+/7d1b3a689be18a48f269c915ee49c19c204c5238
This is the last commit before the reformat changes.

2) run the reformat scripts in git_main external/deqp repo
and land the reformat change.

3) Merge the reformat changes in remote tracking branch
goog/upstream-vulkan-cts-1.3.8 with "-s ours" option.

This CL also cherry-picked this change https://gerrit.khronos.org/c/vk-gl-cts/+/14765,
to address a test that is affected by the reformats.

Only code format change is involved. No function changes
and no other tests are affected.

Ignore-Khronos-First: Reformat scripts need to run in repo directly
Test: make cts
Bug: b/327004039
Change-Id: I8a54434540bd7873c1460b8dc6dde4a94a07d0b7