Avoid spamming logs on infra errors
When build fails due to an infra error (e.g. a repo wasn't cloned),
hundreds of thousands of lines in the form "Case ... was never used!"
messages are printed.  This change limits the log spam to a 100 cases,
followed by a message saying that the list is truncated.

This makes the log files much more manageable.

Components: Framework
VK-GL-CTS issue: 4652

Change-Id: Icb74f2a287fd1d95a4aee161b9ce747c4afeda82