Update vulkan-docs to vk 1.4
This change updates gen_framework script to add support for vk1.4.
Please note that stripNonmatchingAPIs no longer could be used because
we need information from feature section for vk1.4 to properly
generate inl files for Vulkan SC.
With changes to gen_framework also improvement was added that tries
to name getters for properties and features in the same way for
Vulkan and Vulkan SC even if specified extension was promoted in Vulkan
but not in Vulkan SC. This enabled possibility to remove many ifdefs
from CTS code.
This change also updates json files for extensions that were added
to core and adds tests for VkPhysicalDeviceVulkan14Features and
Properties.

VK-GL-CTS issue: 5172

Components: Framework, Vulkan, Vulkan SC

Affects:
dEQP-VK.*

Change-Id: I176edc7468d47babd2f23b7755a0665162113cfb