Bind pipeline between secondary draw and primary draw in nested command buffer KHR tests
In the current CTS for mixed inline and secondary buffer, after a cmdExcuteCommands(secondary command buffer(s) are recorded to execute on a primary command buffer), according to the spec of the command buffers chapter, all state of the primary command buffer is undefined and it will result in invalid draws after cmdExecuteCommands.
To rebind a pipeline between cmdExcuteComands and cmdDraws will lead a meaningful draw and make the test valid.

Components: Vulkan

VK-GL-CTS issue: 5202

Affected tests:
dEQP-VK.*.nested_command_buffers.*

Change-Id: I912835822690b802d22d78eb9299ddbb9b4bc1cb