RESTRICT AUTOMERGE: Introduce incremental-deqp-baseline testlists
For the initial dependency collection pass in incremental-deqp we need
a testlist that is more complete than the incremental-deqp testlist, but
still applicable to every device from GRF perspective.

The incremental deqp infrastructure also doesn't understand modern split
testlists, so this needs to be a single flat list.

Initially these lists are based on the 2019 vulkan lists, and the 2020
egl/gles lists. The difference is due to frozen testlist support being
rolled out a year later for egl/gles, so 2020 is the first frozen year
available.

Adjust filegroup definition for incremental deqp testlists (which need
to be accessible to the incremental deqp target preparer) to include all
the baseline testlists as well.

Bug: b/353581414
Ignore-Khronos-First: Android specific
Change-Id: I6ddec45d3e33e7e38d74ff42534bf3c936f74dae