Rename video decode tests to extract codec name as new group
With the upcoming introduction of AV1 encode tests, we want to
homogenize the nomenclature of both encode and decode tests in
the form:

dEQP-VK.video.(encode/decode).<codec>.*

where <codec> can be h264, h265 or av1 (when they are added).
This is the decode counterpart of CL 15007.

Affects: dEQP-VK.video.decode.*
Components: Vulkan

Change-Id: I92a69bfe64d207e77a3f24a8490359851b4de263