Add extra fast-math decoration in OpPhi tests
In the code
    %comp               = OpFOrdGreaterThan %type_bool %arg1 %arg2
the test was applying the correct fast-math decorations to %comp but not
to %arg1 or %arg2. This meant that the compiler was still able to infer
that the inputs to the comparison could not be certain values and
optimise the results accordingly, causing the test to fail. Add the
decorations to %arg1 and %arg2 as well.

Components: Vulkan
Affects: dEQP-VK.spirv_assembly.instruction.*.float_controls2.*.phi_*

Change-Id: I814f379599be167a32ee081f65b7d824ed5929de