Support standalone executable builds on Android
Only building dEQP as an app on Android is supported. This change allows
building dEQP as a standalone executable on Android that can be run from
adb shell. This is more convenient for iterative testing during driver
development as the command and corresponding output are in one place.

VK-GL-CTS issue: 5320

Components: Framework

Change-Id: Ibeac5160e058b96f9ece30e0eafb991bef570164