Fix error in test logging in all CtsDeqpTestCases
Currently, the fact a test is running is only reported when it has
actually completed.

This also breaks the duration reported, which is always reported as 0ms
or 1ms (the delta between calls to testStarted and testEnded on the
test runner).

I checked the case where tests were failed, skipped or timed out,
everything still worked as expected.

Components: Vulkan
VK-GL-CTS issue: 5323
Affected tests: all CtsDeqpTestCases

Change-Id: I45b702fe811a18b8ea8c0acd8c0e3c632a9313fa