Revert "Fix crash in local read max input attachments test"
This reverts commit 6c3c2ff18c2a80145d443da4ecd1f986f58a44dd. Setting
pColorAttachmentInputIndices to NULL with a non-zero
colorAttachmentCount is an entirely valid thing to do per the spec, that
was what was intended to happen, and any crash is a driver bug. Also, by
skipping setting the depth/stencil mapping this broke the input
attachment indices.

Components: Vulkan

VK-GL-CTS issue: 5253

Affected tests:
dEQP-VK.dynamic_rendering.*.local_read.*

Change-Id: I9c60aca80bc8bc00fc7bb90a832bb8fce6428d04