Fixes for mandatory feature and feature_consistency tests
These tests had become quite confusingly entangled, with the consistency
tests also checking whether various things were supported when this
isn't really necessary. Refactor so that the mandatory_features test
checks whether features are supported and the consistency tests check
that all the structs report this consistently.

Multiview is removed from the mandatory features test for vk1.2 because
it was already mandatory and tested in 1.1. The 1.1 core feature struct
(which was added in 1.2) was already being checked for consistency with
this.

All manadatory feature checking is removed from the 1.2 and 1.3
consistency tests except where the original extensions don't have a
feature struct to check consistency with. (I think this was the original
intent of the tests). The mandatory features test was already correcting
checking whether the features were supported or not.

Add one consistency check that had been missed out from the vk1.3
struct.

Components: Vulkan
Affects: dEQP-VK.info.device_mandatory_features
         dEQP-VK.api.info.vulkan1p*.feature_extensions_consistency

Change-Id: I022ca85e5f1ca1918c9d8ba1e48b6d586afe131a