Clear 2d_view_of_3d images on the same command buffer as the test
Previously, in the sparse versions of these tests, the command buffer
submitted by clearColorImage() wasn't wating on the sparse bind leading
to GPU faults if the graphics queue lost the race with the sparse
binding queue.  Unfortunately, the clearColorImage() helper has no way
to wait on a semaphore so we need to either re-plumb the helper, do a
CPU wait on the sparse bind, or just manually clear the image on the
same command buffer as the rest of the test.  I chose the later option.

Affects:
VK.pipeline.*.image_2d_view_3d_image.*

VK-GL-CTS issue: 5365
Components: Vulkan

Change-Id: I7a5862c8381e832ff9fa3d1835ff0dfdafa7a330