Revert "Fix support check for dynamic state tests"
This reverts commit 05ead5450ec4431373473df21c0998d45ee0c402. The fix
for 4861 leaves blocks of code that can never be executed because the
new if conditions are inside a contradictory 'else' block.

The original code here was intended to enable these features when they
are supported because they were promoted into Vulkan 1.3, which is
supported, but the feature structs are not otherwise returned because of
the partial promotion.

Components: Vulkan
VK-GL-CTS issue: 4861
Affects: dEQP-VK.pipeline.monolithic.extended_dynamic_state.*

Change-Id: Iede354b1888712d9a0701b7ccebcb3b58814f3d6