Replace DE_INLINE with inline
A few functions that access `static` data are explicitly marked as
`static inline`.  Additionally, functions used in C code are also marked
as `static inline` as `inline` does not follow the same rules as in C++.

In some cases `inline` is removed, such as .c files where the function
is defined locally; inlining should be left to the compiler.

Components: Framework
VK-GL-CTS issue: 5409

Change-Id: I4216955420c7290d79a09e0065e7415617cea5e9