Fix OOB VBO reads in dynamic stipple bresenham line tests
These tests use a vertex buffer binding with positions at the start of
the buffer and colors after those positions, so that the color attribute
offset is the amount of position data.

However, a subset of the cases were using a second vertex buffer for an
extra offscreen draw, and that buffer contained a few positions and
colors, with colors stored right after those few positions. This
followed the same principle that positions come first followed by
colors, but does not take into account the color attribute offset.
Instead, positions need to come first and colors need to come second,
but start at the same offset in the buffer as the normal case, so the
attribute offset is correct and matches the pipeline.

Affects:
dEQP-VK.rasterization.primitives.dynamic_stipple_and_topology.bresenham_lines*

Components: Vulkan
VK-GL-CTS issue: 5443

Change-Id: Ib8b2723ae7c74f547297e4d288a5f25f003760be