Workaround multisample issues for local input attachment reads
This is the first change that helps to address the issues relating
to local input attachment reads on multi-sample buffers.

Reading from samples that aren't covered is undefined behaviour.
This change makes the rendered output floating point instead of
integer; this allows a full resolve to be applied rather than
just a sample0 resolve. The shaders are changed to output 1.0 on
success. Only if a sample is covered does the subpassLoad actually
happen. Non-covered samples default to output 1.0.

If any of the covered samples don't load the correct value the
result will be 0.0 and the resolve won't become 1.0.

Affects:
dEQP-VK.*.sampleread.numsamples_*

Components: Vulkan
VK-GL-CTS issue: 5675

Change-Id: Ieef8a6bb4c8eda228602ce78b5e8fc79c583247f