Exclude MAIN_VULKAN_FILTERS from all lists
Exclude the tests listed in MAIN_VULKAN_FILTERS from all test lists.

There is a test list for each year and a device has to pass the
combination of tests lists up to that year. For example a device
released in 2023 needs to pass all the tests until and including
the year 2023 i.e. all the tests from 2019, 2020, 2021, 2022, 2023.
The list of filtered tests needs to be applied retroactively to all
lists, since a "bad" test may be included in a previous year's test
list, if the determination to skip the test was made after it was
already included.

Bug: 407794022
Test: python3 scripts/check_build_sanity.py -r android-mustpass
Ignore-Khronos-First: Android-specific testlist maintenance
Flag: TEST_ONLY
Change-Id: I7a82d8dd4e5cdd28284361c6e481ea226d5dc28d